separate decorators
Put decorator functions into a separate file
