+++ rac_aspace/data_helpers.py
@@ -1,15 +1,7 @@
 import raw_input
 
 from fuzzywuzzy import fuzz
-
-
-def check_dictionary(func):
-    def inner(*args, **kwargs):
-        if not isinstance(args, dict):
-            print('Parameter is not a dictionary')
-            return
-        return func(*args, **kwargs)
-    return inner
+from decorators import check_dictionary
 
 
 @check_dictionary
