+++ rac_aspace/decorators.py
@@ -1,0 +1,9 @@
+
+
+def check_dictionary(func):
+    def inner(*args, **kwargs):
+        if not isinstance(args, dict):
+            print('Parameter is not a dictionary')
+            return
+        return func(*args, **kwargs)
+    return inner
